Stencil code a la MPI to assess performance of bulkComms
In the previous commit I forgot to mention the new test file that Vass
and me wrote in pair programming session (Vass was typing, of course!:-). 

Brad suggested the idea of implementing stencil codes without relying
on Chapel global-view feature, but having the data manually
distributed in per-local defaultRectangular arrays. Halo communication
are done using array slice assignments, and therefore putting to work
our bulkCommStride routines. That way, is similar to an MPI way of
coding, but it ended up being much more productive than the real MPI
version. For instance, we have chosen a simple heat equation as the
stencil code, similar to the one shown here:

https://computing.llnl.gov/tutorials/parallel_comp/#ExamplesHeat

In that page, the MPI version is available and has 230 LOC while our
Chapel version only has 100.

The committed version is just the starting point one (v1). Further
improvements can be done:

1.- Vass suggests that if verification is desired for large problem
sizes, it makes sense to make the reference array
Block-distributed. For small problem sizes DR for the reference array
is probably faster. Also, given this verification, you really only
need to run 3 computation steps to ensure correctness.

2.- Vass also wants to reduce the number of communications needed to
repetitively access the pointers to remote Halos. He proposes to stash
away a._value for the corresponding slices and he points out that "When
stashing away a._value for a slice, need to bump its reference
counter. Since the stashes are going to be pointers, the nil pointer can
indicate the absence of that particular neighbor."

3.- I also think that we can overlap computations and communications
by using Chapel's begin statement. This is because the computation of
the regions that have to be sent to the neighbors together with the
actual communication can be done in parallel with the computation of
the inner part of the array that are not needed by the neighbors locales.



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/bulkComms@20426 3a8e244f-b0f2-452b-bcba-4c88e055c3ca