Factor and rearrange module code, so here.getSubLoc(__primitive("_get_subloc_id")).initTask() can be resolved the first time an "on" clause appears.  Since the type of here (parameterless function) depends on _here and _here is initialized in chpl__init_ChapelLocale, this means that the ChapelLocale module has to be initialized before any use of an "on" clause. DefaultRectangular and DefaultDist use "on" clauses, so this largely entailed factoring the ChapelRootLocale module out of ChapelLocale, and moving the "use DefaultRectangular" statement into ChapelRootLocale. Also, syncvars were factored out of ChapelBase into a new ChapelSyncvar module. The syncvar implementation uses "on" clauses, so ChapelBase had to be split to permit initialization of ChapelLocale to precede the initialization of ChapelSyncvar -- though presumably at least some of the definitions in ChapelBase are needed early.
DETAILS
-------
    modules/internal/ChapelArray.chpl
    modules/internal/DefaultRectangular.chpl
config param noRefCount was moved out of ChapelArray and into
DefaultRectangular.  My changes to the module resolution and initialization
order force DefaultRectangular to be resolved before ChapelArray (in contrast
to the existing order), so its type is undefined unless I move it closer to the
location of its first use.

    modules/internal/ChapelLocale.chpl:30
A dummy initializer for _here was required, since at runtime there are
references to "here" which precede the initialization of rootLocale.  It might
be possible to reorder the modules so that rootLocale is initialized early
enough, but this also works....

    compiler/include/primitive.h
    compiler/AST/primitive.cpp
Added and implemented _get_subloc_id.

    compiler/passes/parallel.cpp
Regularized the if ... else if ... sieve in the main loop in parallel().


git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/caseyb@20456 3a8e244f-b0f2-452b-bcba-4c88e055c3ca