Change the criteria for deciding that a class is unused
A class is considered unused if after function resolution:
  * its default constructor and default type constructor are unresolved
  * all of its subclasses are also unused

Previously, a class was considered unused if after function resolution:
  * its default constructor and default type constructor are unresolved
  * it has at least one field that is not a param, type, or the super field

This fixes the occasional chpldoc.chpl failures. In that test, the super class
Foo has a field, so was deleted as unused.  The subclass has no fields, so was
not deleted as unused. Later, we look at the subclass super field's type which
is now garbage.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20464 3a8e244f-b0f2-452b-bcba-4c88e055c3ca