bulkCommsStride after code review with Vass (part II)
Cleaning old comments and adding new ones. Some of the comments make a
reference to the test example and the line number in which an issue
arise if the corresponding lines of code are deleted.

In BlockDist.chpl
   - old functions corr_direct and corr_inverse have been deleted
      because the new bulkCommConvertCoordinates does the same
      function better.
   - BlockArr.doiBulkTransferStride updated to use the new
       bulkCommConvertCoorindate and tuplify functions. Still pending
       the change of tuplify renamed to bulkCommTuplify(arg).

In DefaultRectangular.chpl

   - All "for in in 1..rank" changed by "for param i in 1..rank"
   - Function renaming:
      - distance --> checkStrideDistance
      - getStrideLevels --> computeBulkStrideLevels 
      - getCount --> computeBulkCount
      - getStride --> computeBulkStride
   - The argument rankcomp[] of these previous functions is now a
       tuple. Also these functions have been simplified and commented.

   - Still pending the movement of proc
     DefaultRectangularArr.doiBulkTransferStride(Barg) where
     Barg._value.isBlockDist() from DefaultRectangular.chpl to
     BlockDist.chpl. 

   - Some variables renaming. For example:
         srcCompRank --> srcWholeDim
         dstCompRank --> dstWholeDim
which are arrays of bools indicating if each corresponding dimension
is accessed as a whole. For example, var A: [1..10,1..10] real. If we
do A[1..10,1..5]=B[..], then dstWholeDim=[true,false].
Probably better names are desirable. 

Minor changes in some test files.

runtime/src/comm/none/comm-none.c still needs to be reimplemented. Now
it only compiles for CHPL_DEVELOPER=0

Alberto just sent me the patch I'm committing, but I did not have time to
do a full review of his changes. So I'll do the revision ASAP and commit again
tomorrow.

For example we need to review the condition of line 1027 of
DefaultRectangular.chpl ,that prevent stridelevel to be 0, which is
not convincing.



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/bulkComms@20468 3a8e244f-b0f2-452b-bcba-4c88e055c3ca