Wow!!!... There was a bug in chpl_comm_gets/puts in comm-gasnet.c.
Bulk communications for arrays slices in which there is no strides
were managed by the doiBulkComm function (the non strided one). In
those cases stridelevels=0 (number of different strides) and we never
checked what if stridelevel=0 for doiBulkCommStride. When I did it, I
found that my chpl_comm_gets/puts functions were not prepared for
that!!. Now they are.

Now the funny condition of line 1027 in DefaultRectangular.chpl, that
was preventing stridelevel to be 0 can be deleted.

A new test for this case also included.

When stridelevels=0, the GASNet strided put/get
functions just do a contiguous copy of count[0] bytes and the
srctride/dststride arrays are ignored (so it basically does a
non-strided put/get). This makes me think that we can probably do away
with the non-strided doiBulkComm function (that now is only used for a
single contiguous comm). 



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/bulkComms@20484 3a8e244f-b0f2-452b-bcba-4c88e055c3ca