Added chpl__tuplify(x), which will be used by the bulkComms branch.
This function wraps the argument in a 1-tuple, if it is not a tuple already.

While it has the following uses, we consider it to be a workaround
for certain language issues, and so not deserving to be a public
library method:

* In BlockCyclic, when rank=1: the 'startIdx' constructor arg is a
scalar vs. the 'lowIdx' field is a tuple.
Ditto for blocksize (constructor arg vs. field).

* When writing rank-independent code for DSI: the results of the
following functions need to be tuple-adjusted:

dsiLow
dsiHigh
dsiAlignedLow
dsiAlignedHigh
dsiStride
dsiAlignment
dsiFirst
dsiLast

* AMR code has 4 uses of tuplify() (which is a different name for the
same function). 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20496 3a8e244f-b0f2-452b-bcba-4c88e055c3ca