Fixing and improving chpl_comm_gets/puts in comm-none.c.
These functions were not compiling for CHPL_DEVELOPER=1 due to 
pointer arithmetic with void* and also because mixed
var declarations and code. This has been solved.

Besides, the previous version of these functions were generic for a
given stridelevels value. Following Vass suggestion I've specialized
the cases for stridelevels=0, 1, 2, or 3, by using a switch construct.  
The original code is still in the default case, so it deals with
stridelevels>3. After a more exhaustive testing of the general case
for stridelevels=4 and 5, we will also specialize for these two values. 



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/bulkComms@20498 3a8e244f-b0f2-452b-bcba-4c88e055c3ca