Adding a few performance optimizations to LULESH
The main one is to change a global 4x8 'const' array from an array to
a tuple of tuples which results in a significant improvement in speed
(embarrassingly large).  I haven't fully diagnosed the root cause yet,
but am guessing that the tuple solution results in replication of the
values whereas the array solution does not (?).

The other improvement is minor, but may have bigger impact at scale:
When we changed from sync var-based arrays to atomic-based, an
array initialization was serialized and localized rather than using
a forall.  Oops.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20504 3a8e244f-b0f2-452b-bcba-4c88e055c3ca