Fix for qthreads build issue
With the recent codegen refactoring, the qthreads build broke due to
the header files it includes, the things they include, and the search
paths provided to qthreads.  The root cause was due to converting
macros to inline functions (required by LLVM) which it turn requires
prototypes more actively (whereas with a macro they only have to be
provided in the use contexts).

This commit fixes that change by moving the inlined funtions into
chpl-comm-compiler-macros.h (now arguably badly named) out of chplrt.h,
the rationale being that these functions are only used by the generated
code (stdchpl.h) not by the runtime (chplrt.h).

While here, I added a few requied #includes that had previously been
handled by virtue of the order in which things happened to be
#included.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20512 3a8e244f-b0f2-452b-bcba-4c88e055c3ca