LULESH variation for study
I was getting a headache keeping all my "not yet ready for prime-time"
LULESH changes straight, so am checking in a working copy here for
the time being, until they can be propagated to the release.  This is
also likely to be the copy that LLNL uses for their LULESH paper.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20514 3a8e244f-b0f2-452b-bcba-4c88e055c3ca