Changed the 'should sync arrays be allowed' future to answer in the negative.
This is our current thinking, and Brad's rationale for it: Imagine we
allowed 'sync' on aggregate types. Then, when is the full/empty bit
set for a multiplicity of values?  When accessing any of them?  Only
when accessing one of them?  How is the bookkeeping done?  Combined
with a lack of compelling use cases.

It is still a future because the compiler does not implement it.
I.e. the compiler accepts sync arrays (and perhaps other aggregate
types) without reporting an error.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20535 3a8e244f-b0f2-452b-bcba-4c88e055c3ca