Commented out debugging calls to writeln() in proc =(a:[],b) and callees.
The issue arises when there is a user module called 'writeln',
e.g. when compiling a file writeln.chpl (even if the file is empty).

I did not investigate in detail, but it seems like that function
is invoked pretty early during normalization, before the proper
definitions of writeln() are encountered. So 'writeln' gets resolved
to a module name, rather than a function name, causing these errors:

$CHPL_HOME/modules/internal/ChapelArray.chpl:1556: In function 'chpl__useBulkTransfer':
$CHPL_HOME/modules/internal/ChapelArray.chpl:1557: error: illegal use of module 'writeln'
$CHPL_HOME/modules/internal/ChapelArray.chpl:1574: In function 'chpl__useBulkTransferStride':
$CHPL_HOME/modules/internal/ChapelArray.chpl:1575: error: illegal use of module 'writeln'
$CHPL_HOME/modules/internal/ChapelArray.chpl:1615: error: illegal use of module 'writeln'

With these calls to writeln() commented out, the issue does not arise.


git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/bulkComms@20539 3a8e244f-b0f2-452b-bcba-4c88e055c3ca