Reparing test regression after codegen refactor.
These tests were failing with --no-local because they
were running out of stack space. Most likely, the use
of C99 anonymous stuctures occupies stack space when
the C compiler is not optimizing. Increasing the amount
of stack space provided to these tests solves the problem.
Note that they require significantly less stack space
if compiled with --fast, and that other than the C99 anonymous
structures, the generated stack variables are the
same (in QuickSort, where the problem was) before
and after the codegen patch.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20544 3a8e244f-b0f2-452b-bcba-4c88e055c3ca