Enabling MassiveThreads support
[reviewed by greg]

This commit adds tasklayer code for MassiveThreads to the runtime/
directory tree.  Most of this code was adopted from the code in
third-party/massivethreads/massivethreads-0.3beta/chapel-if/1.5
directory developed by Jun Nakashima et al.  The main changes involved
were to switch from using the shorthand "myth" to the longer
"massivethreads" for clarity.  The decision to pull this code out of
the massivethreads package and into the source tree was made in order
to simplify maintainability going forward (ours and Jun's).

This change also updates READMEs that refer to MassiveThreads.  A
few changes still need to be made to README.tasks to discuss what
MassiveThreads does by default w.r.t. threading and callstack size.
I've asked Jun to help with that next week if he has time.

To use MassiveThreads should be as simple as setting CHPL_TASKS to
massivethreads and rebuilding.  As with many of the user-level tasking
layers, the callstack size needs to be increased using the standard
flag on the executable command line for many programs.  We may want to
look at increasing its default in the near future.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20549 3a8e244f-b0f2-452b-bcba-4c88e055c3ca