Getting html-based --docs working a bit more seamlessly
This commit takes Lydia's --docs html mode a bit further by
having the compiler invoke the chpldoc2html script automatically
when creating --docs in creole/html mode.  

It also adds space for the necessary creole helper python modules to
the third-party directory, though the packages themselves are not
added because we have not yet cleared them through the legal team.
Instead, the Makefiles in those directories will use the easy_install
capability to download and install these packages to the third-party
directories if/when a user requests them to be installed via
'make creoleparser genshi'.  In future releases, these scripts should
be bundled with Chapel for simplicity.

The compiler invokes chpldoc2html adding these third-party directories
to PYTHONPATH for the command.  If the command fails, it prints out
a message indicating to the user that they may need to do the 'make'
in the third-party directory to install the scripts.

The other main change to this pass is that I wanted to access the
'make' command corresponding to CHPL_MAKE (or util/chplenv/chplmake),
and found that it was set up in a fairly different way than other
CHPL_* environment variables, so I unified it while I was here.  It's
still a fairly clunky way to set this up, but at least it's
consistently clunky.

Documentation and further improvements to follow (probably after
the feature freeze).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20552 3a8e244f-b0f2-452b-bcba-4c88e055c3ca