Environment variable changes for 1.6.
xt-cle and xe-ce are now deprecated, replaced with cray-xt and cray-xe,
respectively.  To aid users in transitioning, deprecated CHPL_HOST_PLATFORM
values will automically be changed to its new corresponding value.  
When this occurs a warning message will be displayed to the user whenever 
chplenv/printchplenv is called.   In addition, two new platforms are supported 
cray-cascade and cray-xk.  


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20568 3a8e244f-b0f2-452b-bcba-4c88e055c3ca