Fix bugs to do with QIO binary I/O and on in writeThis
[Reviewed by Brad]

Previously, if one put an 'on' statement in a writeThis
routine, QIO would explode, since it assumes that the
user-provided writeThis/readThis routines do not change
locale (and that they are run on the same locale as the channel).

Also, if e.g. kind=native is specified when creating a channel,
the dynamic style stored in the channel should record that it
is a binary channel (for slow-path cases; normally the kind=native
will cause binary I/O to be optimized in at compile-time).

Lastly, it was pointed out that whole-array-writing puts spaces
even when doing binary I/O; we just omit the spaces if the
I/O is set to be binary (for some distributions at least).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20578 3a8e244f-b0f2-452b-bcba-4c88e055c3ca