Add the ref intent to the spec
Added the ref intent to the spec. In most places (all, I think) that mention
the out and inout intents, added ref to the list.

Corrected the out intent section to allow the formal to be generic, and take
its type from the type of the actual.

Ran a spell-checker on the files I touched and fixed some spelling typos.

[Reviewed by hilde]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20587 3a8e244f-b0f2-452b-bcba-4c88e055c3ca