Fixed a regression due to r20578 and added a reminder pseudo-test.
The test distributions/deitz/test_block_representation.chpl generates
a warning with a line number in it. So changes to modules/standard/IO.chpl
right now necessitate updating that test's .good files.
However, that test is not run under the standard parallel testing
(due to a SKIPIF), so that update has been overlooked, more than once
in the past.

The test/io/vass/ddata-printing-warning.chpl *is* run under the standard
parallel testing and so intentionally triggers the same warning.
The corresponding .good contains a reminder pointing to
distributions/deitz/test_block_representation.

The right solution is to fix BlockArr.dsiDisplayRepresentation() so it
does not generate the warning. Or eliminate the line number from .good.
But since we are past feature freeze at the moment, I am
creating the above reminder as a workaround.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20589 3a8e244f-b0f2-452b-bcba-4c88e055c3ca