Revert calls to memmove() back to memcpy()
After doing some performance investigation of r20525, change the call
to memmove() in chpl_gen_comm_get/put() back to using memcpy().  These
calls are used for all single word gets/puts, and as was seen in my
initial experimentation, memmove() was worse than memcpy() for moves
of regions < 4K words.

The initial change was made due to valgrind errors where the src and
dest were the same, so we aren't failing any tests (I have on my plate
to investigate why we are generating such code).  We currently do not
generate code where src and dest regions overlap (that we know of),
though we may soon.

[reviewed by gbt, discussed with all]



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20593 3a8e244f-b0f2-452b-bcba-4c88e055c3ca