Changing uses of 'inout' intents to 'ref'
I updated the README.extern file to indicate that 'ref' can/should be
used (rather than 'inout') to pass things by pointer.  Updated some
tests that used 'inout' to verify that this worked as expected, and it
seems to.  Other tests remain to be updated, but these were the only
ones I had time to get to tonight (and the others aren't incorrect
per se, just not the preferred style).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20601 3a8e244f-b0f2-452b-bcba-4c88e055c3ca