Simplify the emacs directory structure to avoid duplication
At the time that we introduced the emacs mode to Chapel, it had just
undergone a major change to the cc-mode which caused us to
conservatively put the major and minor emacs number into this path in
case it changed some more.  It's turned out to be quite stable, and so
we've had a lot of duplication as new versions have come out.  To
remove this duplication (and simplify the maintenance going forward),
this commit reduces the major.minor version number to just the major
number.  In the future, if we need to specialize for a minor version
number, we can add a major.minor subdirectory and put two directories
into the search path (major.minor then major-only).

Note that developers/users who have modified their .emacs file to
reflect this path will need to update to reflect this change.  I've
updated the suggested .emacs lines in the README to do so as well.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20609 3a8e244f-b0f2-452b-bcba-4c88e055c3ca