Adding qthreads version 1.8.1 to the repository
The qthreads team has been working to get some improvements done in
time for the release.  This doesn't quite work out of the box (due to
a miscommunication/misunderstanding about how Michael's ccodegen fix
was to be resolved), but should with some subsequent changes that I'm
about to make.  This captures qthreads as it was released.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20613 3a8e244f-b0f2-452b-bcba-4c88e055c3ca