Applying MassiveThreads patch from Jun
This is a patch that Jun sent me which (a) increases the
default stack size to 32k to match other tasking layers
and (b) implements the serial state for massivethreads.
I've turned on Jun's commit privileges today.

I believe this code will not compile with -Werror (just
re-enabled), so will commit a fix for that separately
(I wanted to capture the original patch as it was).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20617 3a8e244f-b0f2-452b-bcba-4c88e055c3ca