Making tasks-massivethreads.c compile under -Werror for gcc
These are some reasonably minor changes to tasks-massivethreads.c
to get it to compile with CHPL_DEVELOPER=true (which turns on
-Werror for gcc).  Mostly, this involved pushing declarations to
the top of compound statement blocks (why does gcc complain about
this when we're using -std=c99 -- isn't it legal now?), in some
cases by pushing get/set serial macros into a new compound
statement.  The one change that gave me a little pause was to
have ns_task_wrapper() return NULL since it is declared to be of
type void* (since that's what massivethreads wants).  I'll ask
Jun to double-check me that this is appropriate.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20619 3a8e244f-b0f2-452b-bcba-4c88e055c3ca