Removed duplicate testing of SSCA2 and added testing of SSCA2 with non-serialized edge generation and graph construction.
* test/studies/ssca2/vass/SSCA2_main.chpl

Removed - this testing is performed by the release version,
which is essentially the same as of r20627.

* test/studies/ssca2/vass/rmatalt.chpl

Moved to test/studies/ssca2/test-rmatalt/reproduc.chpl
and pointed it to the release version for the code.
See r20275 for a description of what it does.
The graph is reproducible.

* test/studies/ssca2/test-rmatalt/nondet.chpl

Tests edge generation and graph construction without forcing
serialization of the two key loops. (See r20266 and r20627
for discussions.)

As a result of nondeterminism, can verify only the neighbor
histogram (I think), and that it runs to completion successfully.

Like reproduc.chpl, goes to the release version for the code.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20628 3a8e244f-b0f2-452b-bcba-4c88e055c3ca