Add op= and swap to the list of overloadable operators in the spec
Mention that compound assignments can be overloaded in the Assignment
Statements section.  Mention that the swap operator can be overloaded in the
The Swap Statement section.  Add both to the list of operators in the Function
and Operator Overloading section.

I didn't include &&= or ||= in the list of overloadable operators because they
are not defined as operators and doing so would break their short-circuiting
evaluation.

[Reviewed by Sung]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20630 3a8e244f-b0f2-452b-bcba-4c88e055c3ca