Added a bug future on initializers for enumerated domains/subdomains having no effect
From the .future:

Apparently an enumerated domain variable starts out full,
whether it has an initializer or not.
If there is an initializer, it adds elements to an already full domain,
ending up still in a full domain, rather than in domain with only
those elements that are listed in the initializer.

While this is consistent in that an initializer is always
treated as a separate assignment, it is unintuitive for the programmer
and deserves a special treatment, or at least a compiler warning
to the effect that the initializer is ignored.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20638 3a8e244f-b0f2-452b-bcba-4c88e055c3ca