This is a patch to Qthreads to remove an earlier patch
When the new codegen back-end broke Qthreads, a proposed patch
was checked into Qthreads, but not Chapel due to wanting (and
eventually finding) a better solution.  This backs out the
qthreads patch and then re-runs autogen.sh (which is how qthreads
gets itself re-ready to configure).

Haven't tried building yet.  That's next.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20641 3a8e244f-b0f2-452b-bcba-4c88e055c3ca