Fix warnings about const char*-to-char* cast.
[reviewed by Brad]

I was seeing compiler warnings (errors) with
gcc (GCC) 4.4.6 20120305 (Red Hat 4.4.6-4)

Cleans up some things that should be const char* anyway,
and replaces an astr that will subsequently be modified
with a strdup. That potentially will leak (future work -
figure out if it does or not - or call astr
on the broken up parts and free the strdup'd one).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20642 3a8e244f-b0f2-452b-bcba-4c88e055c3ca