Fix QIO on Cygwin enough so that fileIO.chpl works
[reviewed by Brad]

This patch got fileIO.chpl to work for me on cygwin,
with uname -a 
CYGWIN_NT-6.1-WOW64 ... 1.7.16(0.262/5/3) 2012-07-20 22:55 i686 Cygwin

It provides a workaround for pread and preadv on Cygwin
returning error code 61 (ENODATA) if they try to read beyond
the end of a file. Arguably it's a Cygwin bug since pread and preadv
are supposed to return 0 in that case... but in case anybody
wanted to run Chapel on Cygwin we have this fix (which adds
a check under an #ifdef __CYGWIN__ )

Also added a comment about why mmap'd file tests probably will
fail on Cygwin - I believe that Windows will not allow a
file to be truncated while a memory map for it is still open.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20643 3a8e244f-b0f2-452b-bcba-4c88e055c3ca