Made the release version of SSCA2 work with VALIDATE_BC=true; switched studies/ssca2/test-rmatalt/nondet.chpl to test with VALIDATE_BC=true.
This required a workaround for the issue where this addition:

  Lcl_Sum_Min_Dist.add(current_distance_c);

done in f3() did not propagate to f3's caller.

As a result of that issue, the value of Sum_Min_Dist ends up being 0
around SSCA2_kernels.chpl:171. That's where one of the checks is
done when VALIDATE_BC=true. Therefore that check fails.

Perhaps Lcl_Sum_Min_Dist should be passed to f3() with a ref intent,
but that currently does not work, seemingly due to an issue in r20485.

While there, removed the assignment Sum_Min_Dist = 1.0, because it
gets overwritten when Sum_Min_Dist is passed in as an 'out' arg to
approximate_betweenness_centrality().

Reviewed by Sung.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20671 3a8e244f-b0f2-452b-bcba-4c88e055c3ca