Streamlined hashing for associative domains over 'imag' indices; corrected an issue in assigning from tuples to irregular domains.
Hashing on 'imag' domains is now done with a direct call to
chpl__defaultHash(imag), rather than via an implicit conversion from
imag to complex and hashing on the complex.

The issue in assigning from tuples to irregular domains was raised in
r20638 and r20639. The corresponding futures are now retired.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20675 3a8e244f-b0f2-452b-bcba-4c88e055c3ca