Support cnselect coremask attribute as a backstop to numcores.
[Reviewed by Sung.]

Some time ago we removed support for the cnselect coremask attribute as
a way to find out the number of cores on a compute locale, because it
was deprecated and the numcores attribute was a better alternative.  But
it turns out that the version of cnselect on Cray XT systems still only
supports coremask.  So, add support for that as a backstop to numcores.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20676 3a8e244f-b0f2-452b-bcba-4c88e055c3ca