Fixes a problem compiling with OS X Lion
[Reviewed by Brad]

r20461 added a #define _POSIX_C_SOURCE 1, but that's
probably the wrong thing to do (generally it should be
e.g. #define _POSIX_C_SOURCE 200112L), and on Mac OS X Lion
defining it to 1 means that we don't get a vsnprintf. Since
sys_basic.h tries to get these things straight, we just
switch to including that file first.

(r20461 needed to do something to get localtime_r to
 be defined).
 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20680 3a8e244f-b0f2-452b-bcba-4c88e055c3ca