Added a feature request on automatic conversion of 'const' to 'param' when possible.
From .future:
When writing generic code, a const may get initialized with a param
expression for some instantiations and with a non-param expressions
for others. I would like a const to be treated as a param if it is
initialized with a param. The purpose is to enable downstream
optimizations in such a case, for example constant folding, param
function invocation, etc. The example does not show such
optimizations, but I have come across cases where they are desired
in range development.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20688 3a8e244f-b0f2-452b-bcba-4c88e055c3ca