Removing this .future
Since Sung updated the syntax, this seems to mostly be passing.
It may not be passing 100% of the time, but testing has been
noisy enough the past few nights that I don't feel confident,
so I'm going to de-futurize it.  If it fails nondeterministically,
then I propose we make it a .notest future to avoid noise in the
testing system.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20690 3a8e244f-b0f2-452b-bcba-4c88e055c3ca