Making the test system skip this future
This future has been flitting between passing and failing since
it was checked in which results in noise in the testing system
without adding any value (testing futures is only productive if
the futures fail consistently).  I'm making it a .notest to
keep it quiet, but leaving it in the testing system since it
represents an issue that's still worth looking at.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20691 3a8e244f-b0f2-452b-bcba-4c88e055c3ca