Refactoring I/O out of main LULESH module
This is a housecleaning change for LULESH.  Previously the code to
initialize the data structures was in-line with the declarations.
In this change, I've factored it all out into helper functions.
My reason for doing so is that when working on the LLNL paper, we
found ourselves wanting to compute the input cube the way the
standard benchmark does rather than reading it in from a file.
So this refactoring sets us up to have a config saying whether to
compute a cube input dataset or to read the dataset from a file.
That's next.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20698 3a8e244f-b0f2-452b-bcba-4c88e055c3ca