Checkpointing the start of a computation of a cube input set
This new luleshCompute.chpl file computes a cube data set for
LULESH using the same interface as the luleshIO.chpl module
I checked in yesterday.  I don't think I've got it quite right
yet, but this is close and work checkpointing.  Then the next
step will be to permit the user to choose between the two
ways of computing a data set using a config.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20702 3a8e244f-b0f2-452b-bcba-4c88e055c3ca