Merged two initialization strategies into one -- luleshInit.chpl
This commit merges the luleshIO.chpl and luleshCompute.chpl changes
from my recent commits into a single file luleshInit.chpl that can
either compute or read from the file.  If the user sets --filename,
it will read from a file; if s/he sets --elemsPerEdge, it will
compute the cube input set; if s/he sets both, s/he'll get a warning;
if s/he sets neither, s/he'll get an error.

I'm testing this on four execopts now:  a 2x2x2 cube in both file
and compute settings (because they're short and test both code
paths), the sedov15oct set we've been using (because it's interesting
and irregular) and the error case mentioned above.  (I decided that
it wasn't all that worthwhile to test the warning case).

Noticed an interesting thing in the testing system which is that if
you have a CATFILES option and put an error case later in your
.execopts options (i.e., one that doesn't create a new version of the
file to be cat'ed), it cat's the file created by the previous test
run.  This could probably be improved by running a CLEANFILES pass
between every .execopts option, but I just worked around this for
the time being by moving my error case first.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20705 3a8e244f-b0f2-452b-bcba-4c88e055c3ca