Changing node sets from non-distributed arrays to sparse
In doing refactoring on initialization, I realized that the node sets
that we're reading into are defaulting to non-distributed dense
arrays.  Since these nodesets are intended to be subsets of the Nodes
domain, using a sparse subdomain seemed better (more natural; more
aligned with the dense Nodes domain).  Of course, until sparse subdomains
are distributed, these will still be localized to locale #0, but in the
long-term this is much better.

I disabled warnings for the time being because I'm getting an array
serialization warning; I'm about to fix that, but wanted to snapshot
this because it's working.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20709 3a8e244f-b0f2-452b-bcba-4c88e055c3ca