Adding/unifying an ability to stop after some number of cycles
This change commits a config 'maxcycles' to both versions of the
code which permits the computation to be halted prematurely after
some number of cycles have executed for the sake of benchmarking.
For the release lulesh, this is a new feature; for the dense version,
the feature existed, but hadn't realized that there was already a
variable 'cycle' that existed for counting the number of steps
taken, so the feature is updated to use that variable.

Removed a blank line from the .compopts file.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20718 3a8e244f-b0f2-452b-bcba-4c88e055c3ca