3-tuple variation on lulesh
This is a variation on lulesh that uses 3-tuples rather than
triples of x,y,z arrays and values.  I thought it would work better
out of the box, but it doesn't seem to be; though it's also suspicious
that small tuples don't seem to be getting scalar replaced as the
compiler's help message seems to indicate.  Need to look into that.

It seems plausible that this might work better in distributed memory
rather than shared by cutting down on communications.  In the meantime,
it's worth snapshotting and keeping alive.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20723 3a8e244f-b0f2-452b-bcba-4c88e055c3ca