Performance improvement to LULESH
LULESH was suffering from the classic performance hit of declaring a
formal type for an array argument, resulting in needless reindexing
overhead.  I removed the formal type which makes the routine more
consistent with the rest of the benchmark anyway (and checked that
no other functions have this issue).

This was the main scalability bottleneck in my experiments
yesterday.  Now I'll start the process again to see what the
next one is.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20735 3a8e244f-b0f2-452b-bcba-4c88e055c3ca