Adding swap overload test and retiring another
Adding a user swap overload test after finding a future from
Michael on this topic and getting his permission to retire
it if we had another one somewhere.  Michael's original future
was a little about overloading swaps, but also a little about
using them as an I/O operator, supporting their ability to be
chained, etc.  Since we now have a swap operator and a new
I/O operator (<~>) to deal with both of his original concerns,
he was OK with retiring the future as long as there was a test
of user-level swap overloads.  It seemed there wasn't, so I
created this quick and cute one.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20747 3a8e244f-b0f2-452b-bcba-4c88e055c3ca