Add comm diags test for HPCC benchmarks based on the release versions
A while back, Greg was playing around with the stream comm counts test
in Steve's directory, and we noted that they don't test the actual
release versions, so I put it on my list of things to do to add such
tests.

This directory copies the release versions of some of the HPCC
benchmarks that we are actively working on, inserts comm diagnostics
calls around the timed regions, and prints out the results.

From the README:

- stream-ep: Sum comms in each iteration (which should equal 0).
  Currently, there is a local block surrounding the loop, so if there
  was any comm, we'd run into problems with the release version.

- stream: Count comms around the timed loop

- ra: Count comms around the timed loop

- ra-atomics: DISABLED due to non-determinism which leads to variable
  number of forks for the atomics (could enable for testing on a
  platform with network atomics)

- hpl: DISABLED due to non-determinism (haven't investigated)





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20758 3a8e244f-b0f2-452b-bcba-4c88e055c3ca